props.mk: add system property to set AudioFlinger client heap size
Playback of high sample rate multichannel clips fails when repeat mode is set to repeat one. This is because the default heap size allocated for AudioFlinger clients is 4MB which is insufficient.

Set the system property ro.af.client_heap_size_kbyte to 7KB. This property is read in AudioFlinger and multiplied by 1024, so effectively the allocated heap size becomes 7MB which is sufficient to play such clips in repeat mode.

Signed-off-by: bilux <i.bilux@gmail.com>