libshims: Add pthread_mutex_destroy hack
In Pie, Google added checks to prevent destroying of the same mutexes
consecutively, and aborts the program on fortify fatal call.

The check is triggered when tearing down the session, which destroys
the same mutex consecutively by adding the destroy instruction in a loop.
This causes the whole daemon to crash and fail.

Workaround the issue by shimming the function; catching destroyed
mutexes and returning -EBUSY without passing on the bad mutexes to libc.

Signed-off-by: bilux <i.bilux@gmail.com>